abstracted player generation fuctionality from team generation in order to allow for the generation of a single player via command.
